+++ src/CmdProgram.jl
@@ -239,7 +239,7 @@ function _run(
         
         @error timestamp() * "ProgramInputValidationError: $(getfield(p, :name)): fail to run validate_inputs (before running the main command)." validation_function=getfield(p, :validate_inputs) command_template=getfield(p, :cmd) run_id inputs outputs _module=nothing _group=nothing _id=nothing _file=nothing _line=nothing
         rethrow(e)
-        return false, outputs
+        # return false, outputs
     end
 
     @label dry_run_start # COV_EXCL_LINE
@@ -257,7 +257,7 @@ function _run(
     catch e
         @error timestamp() * "ProgramPrerequisitesError: $(getfield(p, :name)): fail to run prerequisites (before running the main command)." prerequisites=getfield(p, :prerequisites) command_template=getfield(p, :cmd) run_id inputs outputs _module=nothing _group=nothing _id=nothing _file=nothing _line=nothing
         rethrow(e)
-        return false, outputs
+        # return false, outputs
     end
 
     # run the main command
@@ -269,7 +269,7 @@ function _run(
     catch e
         @error timestamp() * "ProgramRunningError: $(getfield(p, :name)): fail to run the main command." prerequisites=getfield(p, :prerequisites) command_running=cmd run_id inputs outputs _module=nothing _group=nothing _id=nothing _file=nothing _line=nothing
         rethrow(e)
-        return false, outputs
+        # return false, outputs
     end
 
     # confirmation: validate outputs
@@ -278,7 +278,7 @@ function _run(
     catch e
         @error timestamp() * "ProgramOutputValidationError: $(getfield(p, :name)): fail to run validate_outputs (after running the main command)." validation_function=getfield(p, :validate_outputs) command_running=cmd run_id inputs outputs _module=nothing _group=nothing _id=nothing _file=nothing _line=nothing
         rethrow(e)
-        return false, outputs
+        # return false, outputs
     end
 
     try
@@ -286,7 +286,7 @@ function _run(
     catch e
         @error timestamp() * "ProgramWrapUpError: $(getfield(p, :name)): fail to run wrap_up." wrap_up=getfield(p, :wrap_up) command_running=cmd run_id inputs outputs _module=nothing _group=nothing _id=nothing _file=nothing _line=nothing
         rethrow(e)
-        return false, outputs
+        # return false, outputs
     end
 
     # touch the run_id_file